+++ apps/docs/next.config.mjs
@@ -1,5 +1,8 @@
 /** @type {import('next').NextConfig} */
 const nextConfig = {
+  experimental: {
+    optimizePackageImports: ['@codefast/ui'],
+  },
   images: {
     remotePatterns: [
       {