+++ packages/ui/src/components/radio.tsx
@@ -10,7 +10,6 @@ import { cn } from '@/lib/utils';
  * -------------------------------------------------------------------------- */
 
 type RadioElement = HTMLInputElement;
-
 interface RadioProps extends InputHTMLAttributes<HTMLInputElement> {
   onValueChange?: (value: string) => void;
 }